+++ lib/tensorflow/ffi.rb
@@ -62,6 +62,7 @@ module TensorFlow
     attach_function :TFE_OpSetDevice, %i[pointer string pointer], :pointer
     attach_function :TFE_OpGetDevice, %i[pointer pointer], :string
     attach_function :TFE_OpAddInput, %i[pointer pointer pointer], :void
+    attach_function :TFE_OpAddInputList, %i[pointer pointer int pointer], :void
     attach_function :TFE_OpGetAttrType, %i[pointer string pointer pointer], :int
     attach_function :TFE_OpSetAttrString, %i[pointer string pointer size_t], :void
     attach_function :TFE_OpSetAttrInt, %i[pointer string int64_t], :void
@@ -69,6 +70,8 @@ module TensorFlow
     attach_function :TFE_OpSetAttrBool, %i[pointer string uint8], :void
     attach_function :TFE_OpSetAttrType, %i[pointer string int], :void
     attach_function :TFE_OpSetAttrShape, %i[pointer string pointer int pointer], :void
+    attach_function :TFE_OpSetAttrTypeList, %i[pointer string pointer int], :void
+    attach_function :TFE_OpSetAttrShapeList, %i[pointer string pointer pointer int pointer], :void
     attach_function :TFE_Execute, %i[pointer pointer pointer pointer], :pointer
   end
 end