+++ lib/tensorflow.rb
@@ -20,6 +20,7 @@ require "tensorflow/image"
 require "tensorflow/io"
 require "tensorflow/linalg"
 require "tensorflow/math"
+require "tensorflow/nn"
 require "tensorflow/ops"
 require "tensorflow/raw_ops"
 require "tensorflow/strings"
@@ -67,6 +68,7 @@ module TensorFlow
     extend Forwardable
     def_delegators Linalg, :eye, :matmul
     def_delegators Math, :abs, :acos, :acosh, :add, :add_n, :argmax, :argmin, :asin, :asinh, :atan, :atan2, :atanh, :cos, :cosh, :cumsum, :divide, :equal, :exp, :floor, :greater, :greater_equal, :less, :less_equal, :logical_and, :logical_not, :logical_or, :maximum, :minimum, :multiply, :negative, :not_equal, :pow, :reduce_all, :reduce_any, :reduce_logsumexp, :reduce_max, :reduce_mean, :reduce_min, :reduce_prod, :reduce_sum, :round, :scalar_mul, :sigmoid, :sign, :sin, :sinh, :sqrt, :square, :subtract, :tan, :tanh, :truediv
+    def_delegators NN, :space_to_batch
 
     def library_version
       FFI.TF_Version