+++ test/nn_test.rb
@@ -1,0 +1,4 @@
+require_relative "test_helper"
+
+class NNTest < Minitest::Test
+end